fixed all refactoring suggestions
