+++ qbraid_qir/qasm3/profiles/__init__.py
@@ -1,0 +1,18 @@
+# profile/__init__.py
+from .core import (
+    ProfileCapabilities,
+    ProfileRestrictions,
+    Profile,
+    BaseProfile,
+    AdaptiveProfile,
+    ProfileRegistry,
+)
+
+__all__ = [
+    "ProfileCapabilities",
+    "ProfileRestrictions",
+    "Profile",
+    "BaseProfile",
+    "AdaptiveProfile",
+    "ProfileRegistry",
+]
